requirements.txt
# LLM Providers
openai==1.12.0
anthropic==0.18.0

# LangChain
langchain==0.1.10
langchain-community==0.0.25
langchain-openai==0.0.6
langchain-anthropic==0.1.4
langgraph==0.0.30

# Vector Store
chromadb==0.4.22

# Document Processing
pypdf2==3.0.1
pdfplumber==0.10.3
pillow==10.2.0

# Data Processing
pandas==2.2.0
numpy==1.26.4
pydantic==2.6.1
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
requests==2.31.0
aiohttp==3.9.3

# Monitoring & Logging
structlog==24.1.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
